Basic Data Processing Functions
Added functions for basic data processing. This includes finding "NA" times in list of times, converting a DataFrame or Array{Any} containing strings that can be converted to Float32 to Array{Float32}, finding expression data for a list of relevant genes that have a minimum average expression and a coefficient of variation within a given range.