Added Training Functions.
Addedt training functions, such as a custom train! function, mytrain!, and a custom @epochs macro, @myepochs. mytrain! stores the loss over the entire epoch and returns the average loss for that epoch. @myepochs has been modified to store the output of mytrain!. Additionally, @myepochs allows a training to be ended before reaching the maximum (user defined) number of epochs iff loss increases for a minimum (user defined) number of epochs.

Co-Authored-By: nicolasbarra <nicolasbarra@users.noreply.github.com>